handle more errors, MD5, Complete, eTag

+++ src/main/java/com/cs/muic/backend/SimpleObject/BucketController.java
@@ -2,9 +2,12 @@ package com.cs.muic.backend.SimpleObject;
 
 
 import com.mongodb.operation.BatchCursor;
+import com.oracle.javafx.jmx.json.impl.JSONMessages;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
 import org.springframework.data.domain.Range;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.DigestUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.ServletException;
@@ -14,8 +17,9 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.Part;
 import java.io.*;
 import java.net.SocketTimeoutException;
-import java.util.Arrays;
-import java.util.List;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -32,13 +36,19 @@ public class BucketController {
     }
 
     @PostMapping(value = "/{bucketName}", params = "create")
-    Bucket newBucket(@PathVariable String bucketName, HttpServletResponse response){
-        Bucket bucket = new Bucket(bucketName);
-//        bucket.setCreated(bucket.getId().getTimestamp());
-//        bucket.setModified(bucket.getId().getTimestamp());
-        new File("./buckets/"+bucketName).mkdirs();
-        Bucket resBucket = repository.save(bucket);
-        return resBucket;
+    ResponseEntity newBucket(@PathVariable String bucketName, HttpServletResponse response){
+        if (repository.existsBucketByName(bucketName)){
+            return ResponseEntity.badRequest().build();
+        }
+        try{
+            Bucket bucket = new Bucket(bucketName);
+            new File("./buckets/"+bucketName).mkdirs();
+            Bucket resBucket = repository.save(bucket);
+            return ResponseEntity.ok(resBucket);
+        }
+        catch (Exception e){
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @GetMapping("/buckets/{id}")
@@ -52,20 +62,37 @@ public class BucketController {
     }
 
     @DeleteMapping(value = "/{bucketname}", params = "delete")
-    void delete(@PathVariable String bucketname){
+    ResponseEntity delete(@PathVariable String bucketname){
         Bucket B = repository.findBucketByName(bucketname);
-        repository.delete(B);
+        if (B==null){
+            return ResponseEntity.badRequest().build();
+        }
+
+        if (new File("./buckets/"+bucketname).delete()){
+            repository.delete(B);
+            return ResponseEntity.ok().build();
+        }
+        else{
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @GetMapping(value = "/{bucketname}", params="list")
-    Bucket getObjects(@PathVariable String bucketname){
+    ResponseEntity getObjects(@PathVariable String bucketname){
         Bucket B = repository.findBucketByName(bucketname);
-        return B;
+        if (B == null){
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok(B);
     }
 
     @PostMapping(value = "/{bucketName}/{objectName}", params = "create")
-    void createObject(@PathVariable String bucketName, @PathVariable String objectName){
+    ResponseEntity createObject(@PathVariable String bucketName, @PathVariable String objectName){
         Bucket B = repository.findBucketByName(bucketName);
+        if (B == null){
+            return ResponseEntity.badRequest().build();
+        }
+
         String objectKey = objectName.replace('.', '/');
 
         if (!B.objects.containsKey(objectKey)){
@@ -73,8 +100,15 @@ public class BucketController {
 
             B.objects.put(objectKey, newObj);
             //TODO: Add object
+            B.setModified();
+            repository.save(B);
+            return ResponseEntity.ok().build();
         }
-        repository.save(B);
+
+        else{
+            return ResponseEntity.badRequest().build();
+        }
+
     }
 
 //    private boolean contains(List<Content> objs, String objectName){
@@ -88,12 +122,12 @@ public class BucketController {
 //    }
 
     @PutMapping(value = "/{bucketName}/{objectName}",params = "partNumber")
-    public ResponseEntity UploadAllParts(@PathVariable("bucketName") String bucketName,
-                                         @PathVariable("objectName") String objectName,
-                                         @RequestParam("partNumber") String partNumber,
-                                         @RequestHeader("Content-Length") String contentLength,
-                                         @RequestHeader("Content-MD5") String contentMD5,
-                                         HttpServletRequest request){
+    ResponseEntity UploadAllParts(@PathVariable("bucketName") String bucketName,
+                                 @PathVariable("objectName") String objectName,
+                                 @RequestParam("partNumber") String partNumber,
+                                 @RequestHeader("Content-Length") String contentLength,
+                                 @RequestHeader("Content-MD5") String contentMD5,
+                                 HttpServletRequest request){
 
         String objectKey = objectName.replace('.', '/');
         if (repository.findBucketByName(bucketName).objects.containsKey(objectKey)){
@@ -106,12 +140,45 @@ public class BucketController {
 
                 FileUtils.copyInputStreamToFile(inputStream, target);
 
+
+
+                FileInputStream fileInputStream = new FileInputStream(target);
+                String md5 = DigestUtils.md5DigestAsHex(fileInputStream);
+                String targetLen = Long.toString(target.length());
+                fileInputStream.close();
+
+                HashMap<String, String> json = new HashMap<>();
+                json.put("md5", md5);
+                json.put("length", targetLen);
+                json.put("partNumber", partNumber);
+
+
+                boolean md5Match = md5.equals(contentMD5);
+                boolean lenMatch = targetLen.equals(contentLength);
+
+                if (md5Match && lenMatch){
+                    System.out.println("200 OK");
+
+                    return ResponseEntity.ok(json);
+                }
+                else if (!lenMatch){
+                    System.out.println("Content-Length not matched");
+                    target.delete();
+                    json.put("error", "LengthMismatched");
+                    return ResponseEntity.badRequest().body(json);
+                }
+                else{
+                    System.out.println("MD5 not matched");
+                    target.delete();
+                    json.put("error", "MD5Mismatched");
+                    return ResponseEntity.badRequest().body(json);
+                }
+
             }catch (IOException e){
                 System.out.println("+++++++++++++failed to process request++++++++++++");
-                return ResponseEntity.badRequest().build();
             }
 
-            return ResponseEntity.ok().build();
+            return ResponseEntity.badRequest().build();
         }
 
         else{
@@ -122,35 +189,65 @@ public class BucketController {
 
 // ------- TODO havent done MD5 -------------
     @PostMapping(value = "/{bucketName}/{objectName}",params = "complete")
-    public ResponseEntity completeMultiPartUpload(@PathVariable("bucketName") String bucketName,
-                                                @PathVariable("objectName") String objectName){
+    ResponseEntity completeMultiPartUpload(@PathVariable("bucketName") String bucketName,
+                                            @PathVariable("objectName") String objectName){
+        HashMap<String, String> json = new HashMap<>();
+        json.put("name", objectName);
+
+        Bucket bucket = repository.findBucketByName(bucketName);
+        if (bucket  == null){
+            json.put("error", "InvalidBucket");
+            return ResponseEntity.badRequest().body(json);
+        }
+
+        String objectKey = objectName.replace('.', '/');
+        Content object = bucket.objects.get(objectKey);
+        if (object == null){
+            json.put("error", "InvalidObjectName");
+            return ResponseEntity.badRequest().body(json);
+        }
         File directory = new File("./buckets/"+bucketName);
         File[] filesInDir = directory.listFiles();
         List<File> parts = Arrays.stream(filesInDir)
                 .filter(f -> f.getName().split(".part")[0].equals(objectName))
                 .collect(Collectors.toList());
+        Collections.sort(parts);
         long contentLength = 0;
         if (!parts.isEmpty()){
-            for (File part : parts){
-                contentLength+=part.length();
+            try{
+                MessageDigest md5 = MessageDigest.getInstance("MD5");
+                ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+                for (File part : parts){
+                    contentLength+=part.length();
+                    byte[] partMD5 = md5.digest(IOUtils
+                            .toByteArray(new FileInputStream(part)));
+                    outputStream.write(partMD5);
+                    object.partMD5s.put(part.getName().replace('.', '/'), partMD5);
+                }
+                byte[] fullMD5 = md5.digest(outputStream.toByteArray());
+                String eTag = Base64.getEncoder().encodeToString(fullMD5)
+                        +"-"+Integer.toString(parts.size());
+                object.seteTag(eTag);
+                object.setContentLength(contentLength);
+                object.setComplete(true);
+                object.setModified();
+                bucket.objects.put(objectKey, object);
+                bucket.setModified();
+                repository.save(bucket);
+
+                json.put("eTag", eTag);
+                json.put("length", Long.toString(contentLength));
+                return ResponseEntity.ok(json);
+            }
+            catch (NoSuchAlgorithmException | IOException e ){
+                json.put("error", "NoSuchAlgorithmException|IOException");
             }
         }
-
-        Bucket bucket = repository.findBucketByName(bucketName);
-        String objectKey = objectName.replace('.', '/');
-        Content object = bucket.objects.get(objectKey);
-        object.setContentLength(contentLength);
-        object.setComplete(true);
-        bucket.objects.put(objectKey, object);
-        repository.save(bucket);
-
-
-
-        return null;
+        return ResponseEntity.badRequest().body(json);
     }
 
     @DeleteMapping(value = "/{bucketName}/{objectName}",params = "partNumber")
-    public ResponseEntity deletePart(@PathVariable("bucketName") String bucketName,
+    ResponseEntity deletePart(@PathVariable("bucketName") String bucketName,
                                     @PathVariable("objectName") String objectName,
                                     @RequestParam("partNumber") String partNumber){
         File file = new File("./buckets/"+bucketName+"/"+objectName+".part"+partNumber);
@@ -164,7 +261,7 @@ public class BucketController {
     }
 
     @DeleteMapping(value = "/{bucketName}/{objectName}",params = "delete")
-    public ResponseEntity deleteObject(@PathVariable("bucketName") String bucketName,
+    ResponseEntity deleteObject(@PathVariable("bucketName") String bucketName,
                                        @PathVariable("objectName") String objectName){
         File directory = new File("./buckets/"+bucketName);
         File[] filesInDir = directory.listFiles();
@@ -179,6 +276,7 @@ public class BucketController {
             String objectKey = objectName.replace('.', '/');
             Bucket bucket = repository.findBucketByName(bucketName);
             bucket.objects.remove(objectKey);
+            bucket.setModified();
             repository.save(bucket);
         }
         return null;
@@ -201,8 +299,9 @@ public class BucketController {
 //            String[] rs = ranges.substring(6).split(",");
 //            for (String range : rs){
                 String range = ranges.substring(6);
-                Pattern pattern = Pattern.compile("(-?\\d*)-(-?\\d*)");
+                Pattern pattern = Pattern.compile("(-?[0-9]*)-(-?[0-9]*)");
                 Matcher matcher = pattern.matcher(range);
+                matcher.find();
                 String fromstr = matcher.group(1);
                 String tostr = matcher.group(2);
                 long contentLength = object.getContentLength();
@@ -226,7 +325,7 @@ public class BucketController {
                     to = Long.parseLong(tostr);
                 }
 
-
+                // convert negative to positive
                 if (from < 0){
                     from = from + contentLength +1;
                 }
@@ -234,24 +333,98 @@ public class BucketController {
                     to = to + contentLength + 1;
                 }
 
+                if (from > to || to > contentLength){
+                    return ResponseEntity.badRequest().build();
+                }
+
+                // Find from which part to which part is in range
+                File directory = new File("./buckets/"+bucketName);
+                File[] filesInDir = directory.listFiles();
+                List<File> parts = Arrays.stream(filesInDir)
+                        .filter(f -> f.getName().split(".part")[0].equals(objectName))
+                        .collect(Collectors.toList());
+                Collections.sort(parts);
+                List<InputStream> PartsToBeRead = new ArrayList<>();
+                int firstPart = 0;
+                int lastPart = 0;
+                long startAtFirstPart = 0;
+                long endAtLastPart = 0;
+                long acc = 0;
+                int i = 0;
+                long diffAtEnd = 0;
+                boolean inRange = false;
+                try {
+                    for (File part: parts){
+                        acc+=part.length();
+                        i++;
+                        if (!inRange && from <= acc){
+                            inRange = true;
+                            firstPart = i;
+                            startAtFirstPart = part.length() - (acc - from);
+                            InputStream inputStream = new FileInputStream(part);
+                            if (startAtFirstPart != 0){
+                                inputStream.skip(startAtFirstPart - 1);
+                            }
+                            PartsToBeRead.add(inputStream);
+                            if (to <= acc){
+                                lastPart = i;
+                                endAtLastPart = part.length() - (acc - to);
+                                diffAtEnd = acc - to;
+                                break;
+                            }
+                        }
+
+                        else if (to <= acc){
+                            lastPart = i;
+                            endAtLastPart = part.length() - (acc - to);
+                            diffAtEnd = acc - to;
+                            InputStream inputStream = new FileInputStream(part);
+                            PartsToBeRead.add(inputStream);
+                            break;
+                        }
+
+                        else if (inRange){
+                            InputStream inputStream = new FileInputStream(part);
+                            PartsToBeRead.add(inputStream);
+                        }
+                    }
+                    SequenceInputStream combinedIS = new SequenceInputStream(
+                            Collections.enumeration(PartsToBeRead));
 
 
-//                OutputStream outputStream = response.getOutputStream();
-                byte[] buffer = new byte[2048];
-                int read;
+//                    i = 0;
+//                    acc = 0;
 
+                    OutputStream outputStream = new ByteArrayOutputStream();
+                    byte[] buffer = new byte[2048];
+                    int read;
+//                        for (File part : PartsToBeRead) {
+//                            InputStream inputStream =
+//                        }
+                    long toRead = to - from + 1;
+                    while ((read = combinedIS.read(buffer)) > 0){
+                        if ((toRead -= read) > 0){
+                            outputStream.write(buffer, 0, read);
+                            outputStream.flush();
+                        }
+                        else{
+                            outputStream.write(buffer, 0, (int)toRead+read);
+                            outputStream.flush();
+                            break;
+                        }
+                    }
+//                    IOUtils.copyLarge(outputStream, response.getOutputStream());
+                    return ResponseEntity.ok().build();
 
-
-
-//            }
-
-
+                }
+                catch (IOException e){
+                    return ResponseEntity.notFound().build();
+                }
 
             }
             else {
                 return ResponseEntity.notFound().build();
             }
-            return null;
         }
         else{
             return ResponseEntity.notFound().build();
